refactoring code for week4
