+++ 02-express-tutorial/controllers/people.js
@@ -1,4 +1,5 @@
 const { people } = require('../data');
+
 const addPerson = (req, res) => {
   const { name } = req.body;
   if (!name) {
@@ -31,6 +32,17 @@ const replacePeople = (req, res) => {
   res.status(200).json({ success: true, data: newPeople });
 };
 
-// const updatePerson =
+const deletePerson = (req, res) => {
+  const person = people.find((person) => person.id === Number(req.params.id));
+  if (!person) {
+    return res
+      .status(404)
+      .json({ success: false, msg: `no person with id ${req.params.id}` });
+  }
+  const newPeople = people.filter(
+    (person) => person.id !== Number(req.params.id)
+  );
+  return res.status(200).json({ success: true, data: newPeople });
+};
 
-module.exports = { addPerson, getPeople, replacePeople };
+module.exports = { addPerson, getPeople, replacePeople, deletePerson };
