+++ 02-express-tutorial/routes/people.js
@@ -1,11 +1,11 @@
 const express = require('express');
 const router = express.Router();
-const { people } = require('../data');
+// const { people } = require('../data');
 const {
   addPerson,
   getPeople,
   replacePeople,
-  updatePeople,
+  deletePerson,
 } = require('../controllers/people.js');
 
 router.get('/', getPeople);
@@ -29,16 +29,5 @@ router.post('/', addPerson);
 
 router.put('/:id', replacePeople);
 
-router.delete('/:id', (req, res) => {
-  const person = people.find((person) => person.id === Number(req.params.id));
-  if (!person) {
-    return res
-      .status(404)
-      .json({ success: false, msg: `no person with id ${req.params.id}` });
-  }
-  const newPeople = people.filter(
-    (person) => person.id !== Number(req.params.id)
-  );
-  return res.status(200).json({ success: true, data: newPeople });
-});
+router.delete('/:id', deletePerson);
 module.exports = router;
